Extracted listener for ClickListenerForDatePicker
