+++ app/src/main/java/com/example/habitsmasher/ui/dashboard/EditHabitFragment.java
@@ -21,6 +21,7 @@ import com.example.habitsmasher.DaysTracker;
 import com.example.habitsmasher.Habit;
 import com.example.habitsmasher.R;
 import com.example.habitsmasher.listeners.ClickListenerForCancel;
+import com.example.habitsmasher.listeners.ClickListenerForDatePicker;
 
 /**
  *  UI Class that represents and specifies the behaviour of the dialog
@@ -95,12 +96,7 @@ public class EditHabitFragment extends DialogFragment{
         _dateText.setText(DatePickerDialogFragment.parseDateToString(_editHabit.getDate()));
 
         // listener for the date picker dialog fragment
-        _dateText.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                openDatePickerDialog();
-            }
-        });
+        _dateText.setOnClickListener(new ClickListenerForDatePicker(getFragmentManager(), _dateText));
 
         // when confirm button is clicked
         confirmButton.setOnClickListener(new View.OnClickListener() {
@@ -239,27 +235,5 @@ public class EditHabitFragment extends DialogFragment{
                 Log.d("Tracker Status", _tracker.getDays());
             }
         });
-    }
-
-    /**
-     * Opens the calendar dialog used for date selection
-     */
-    private void openDatePickerDialog() {
-        DatePickerDialogFragment datePickerDialogFragment = new DatePickerDialogFragment(new DatePickerDialog.OnDateSetListener() {
-            /**
-             * Sets the text of the date select view to reflect selected date
-             * @param view
-             * @param year year of selected date
-             * @param month month of selected date (integer from 0 to 11)
-             * @param day day of month of selected date
-             */
-            @Override
-            public void onDateSet(DatePicker view, int year, int month, int day) {
-                int correctedMonth = month + 1;
-                String date = day + "/" + correctedMonth + "/" + year;
-                _dateText.setText(date);
-            }
-        });
-        datePickerDialogFragment.show(getFragmentManager(), "DatePickerDialogFragment");
     }
 }
