+++ app/src/main/java/com/example/habitsmasher/ui/history/EditHabitEventFragment.java
@@ -18,6 +18,7 @@ import com.example.habitsmasher.DatePickerDialogFragment;
 import com.example.habitsmasher.HabitEvent;
 import com.example.habitsmasher.R;
 import com.example.habitsmasher.listeners.ClickListenerForCancel;
+import com.example.habitsmasher.listeners.ClickListenerForDatePicker;
 
 /**
  * The EditHabitEventFragment class
@@ -69,12 +70,7 @@ public class EditHabitEventFragment extends DialogFragment {
         _dateText.setText(DatePickerDialogFragment.parseDateToString(_editHabitEvent.getDate()));
 
         // Add listener to date text to open date picker
-        _dateText.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                openDatePickerDialog();
-            }
-        });
+        _dateText.setOnClickListener(new ClickListenerForDatePicker(getFragmentManager(), _dateText));
 
         // Add listener to confirm button that propagates habit event editing
         confirmButton.setOnClickListener(new View.OnClickListener() {
@@ -102,27 +98,5 @@ public class EditHabitEventFragment extends DialogFragment {
         cancelButton.setOnClickListener(new ClickListenerForCancel(getDialog(), TAG));
 
         return view;
-    }
-
-    /**
-     * Opens the calendar dialog used for date selection
-     */
-    private void openDatePickerDialog() {
-        DatePickerDialogFragment datePickerDialogFragment = new DatePickerDialogFragment(new DatePickerDialog.OnDateSetListener() {
-            /**
-             * Sets the text of the date select view to reflect selected date
-             * @param view
-             * @param year year of selected date
-             * @param month month of selected date (integer from 0 to 11)
-             * @param day day of month of selected date
-             */
-            @Override
-            public void onDateSet(DatePicker view, int year, int month, int day) {
-                int correctedMonth = month + 1;
-                String date = day + "/" + correctedMonth + "/" + year;
-                _dateText.setText(date);
-            }
-        });
-        datePickerDialogFragment.show(getFragmentManager(), "DatePickerDialogFragment");
     }
 }
